Add a class for searching in the Trie struct
This adds a class to handle an in-memory representation of the trie
structure generated by prevodb. The trie can be queried for words
beginning with a prefix. The results are filled into an existing
array.

The class also includes a main function so that it can be compiled
outside of the Android SDK for testing.