pdb-db: Resolve references to an article and mark number
Any 'cel' attributes in the article are now marked in a list instead
of writing out the value to the article buffer. While the article is
saved all of the references are resolved using the hash table so that
it can instead store the article and mark number to make it easier for
the Android application to find the right reference.