pdb-db: Write out a binary file containing the sections of each article
Instead of dumping out the XML and getting the app to do process it,
the plan is now to write out a binary file with a list of
SpannableStrings for the app to display directly. This changes prevodb
to start building the array of strings. The XML document is now parsed
using a PdbDoc instead of trying to parse the data directly from
Expat. This makes the code a lot easier to maintain because it doesn't
have to manage a separate stack for the state and certain branches of
the tree can be handled multiple times.

Although the framework is setup to add spans to the strings all of the
strings are currently just output directly with no spans.

This patch breaks the translation indices and the marks for all of the
sections are zero. These will be fixed in later patches.