search: Add an EditText to the top of the search view
With a soft keyboard it's not possible to make the keyboard appear
unless an EditText is activated so we can't just use the default
filtering mechanism to search. This adds an edit control which sets
the filter text whenever it is changed so that the user has something
to click on to bring up the keyboard.