pdb-db: Split the sences into separate paragraphs
The parser for spannable strings has been split up so that it stores
its state in a separate struct. There is a separate function to handle
parsing a stack entry pointing to a node. The stack entries can now be
one of three types; a node entry, a close span entry and a paragraph
marker entry. The paragraph marker entry is used to break a paragraph
after the tag is closed. The paragraph separators are inserted by
setting a flag on the parser state. The next time any text is added
the paragraph marker will be inserted. That way it won't duplicate
paragraph markers or add redundant ones at the end of the string. The
array of element types now also has an optional callback to do
additional processing for that element. This is used for the 'snc' tag
to start the paragraph and insert the sence number.