ArticleActivity: Handle reference spans
Reference spans are now converted to a ClickableSpan. The onClick
method either launches a new activity with an intent for the given
article and span or just directly scrolls to the right section if it
is within the same article. The section number from the intent when
the activity is started still doesn't cause it to scroll because I
think this would need something to wait until the layouting has
occurred.