pdb-db: Add index entries for the translations
This adds an index entry to the corresponding language for each <trd>
tag. The index entries have been replaced with a union that can either
represent a direct pointer to a section or a mark to indirectly
reference the section. That way the entries for the translations can
be added without knowing the section number.