pdb-revo: Allow the revo zip file to actually be a directory
Instead of requiring the zip file to be passed directly to pdb-revo,
it can now also accept the unzipped directory. This is a lot faster
than launching a separate process to unzip each file.

All of the functions in pdb-revo now have two implementations
depending on whether the data is in a zip file or a directory.

Anything that was using pdb-revo now gives paths relative to the revo/
directory instead of including revo/.