ArticleActivity: Scroll to the right section on initial create
The scroll view for the article is now implemented as a subclass so
that it can override the onLayout function. This is used to delay
scrolling to the right widget until after the initial layout in order
to ensure the views all have the right position.

The article activity now uses this to scroll to the right section on
onCreate.