pdb-db: Search for the translations directly after creating a section
Instead of searching for all of the translations after creating the
article, then are now searched for after creating each section. That
way the translation references can directly point to the right
section. This is necessary because some translations are not under a
section with a mrk attribute.

There are now two ways to search for translations; either recursively
or by just looking at the direct children of a given node. The first
is used when processing an entire drv or a spannable string and the
second is used when processing a container element whose children may
generate other sections.

The references, links and marks have been renamed and reorganised.
There are now comments to explain what they are used for.