Add more versions of the icon with different resolutions
The icon looked a bit naff on the home screen because it was being
scaled up. This adds ldpi, hdpi and xhdpi versions of it so Android
can pick the right size.

The icon ID has been renamed to ic_launcher and the old icon that was
in the stub package has been removed.