pdb-db: Try less precise marks if some are missing
If a link is found to a missing mark, instead of immediately giving up
on the mark it will now try increasingly less precise marks until it
finds a match. The mark is made less precise by removing one of the
components separated by dots. It sounds like some of the articles
deliberately point to a section that is expected to be written later
so this should at least let it find the right article.