pdb-db: Expand abbreviations in 'uzo' tags with tip stl
Any uzo tags that have the tip attribute set to ‘stl’ will now be
expanded instead of showing the abbreviation.

PdbDbRefType has been replaced with PdbReplacement so that the code to
look for the right replacement in the list can be shared.

The element handlers for spannable string elements can now set a flag
on the parse state to ignore the element's children. This is needed to
avoid adding the abbreviation once its expansion has been added.