Add menus to the article and search activities
The article and search activities now have menus to jump back to the
language selection and search activities. There is a common menu class
with static utility methods to implement the menu actions. Whenever a
language is selected, a preference is updated to store the last
language. When the search option is pressed the search for the last
used language is shown.