pdb-db: Temporarily store span offsets as byte offsets
In the article file, the span offsets are stored as a count of UTF-16
code points to make it easier to work with from a Java string.
However, to make it easier to generate a database targetted for other
applications, this patch makes it now store the offsets in memory as
UTF-8 byte counts. They are then only converted to UTF-16 offsets as
the file is written.