Support copying definitions from a context menu
When a long press is performed on the text view for a definition it
now pops up a context menu with an option to copy it to the clipboard.
The definitions are now represented in a tiny subclass of TextView
which stores some extra context menu information to grab the text back
out.

The clipboard is set using a helper class that will either use HTML
data or plain text depending on is supported by the SDK. The
appropriate methods are retrieved using reflection so that it won't
break on older Android.