Add context menu options to create flashcards via Anki
This adds two menu items to the context menu that get shown for each
definition which will create a flashcard either with the word on the
front and the definition on the back or vice versa. The flashcard is
created by just launching the intent
org.openintents.action.CREATE_FLASHCARD. At least Anki handles this
but presumably there may be other flash card apps that can do this
too.