Add an option to change the text size
There is now a menu option in the article view to 'zoom' the text.
This will create and display an instance of ZoomControls with a
timeout so that it will be hidden after 10 seconds of not pressing the
buttons. When the buttons are pressed a configuration option is
updated which causes the text size to change. There are 10 text sizes
which are used to scale the original text size of the layouts.