Add a shared preference for limiting the displayed languages
This adds a shared preference called ‘selectedLanguages’ which is a
comma-separated list of language codes. If this is set then the
article activity will skip sections of the article that have a code
which isn't in the list and the languages activity won't list
the unselected languages.

There is a new class to help parse the comma separated list which
stuffs the language codes into a sorted array so that it can be binary
chopped.

There is currently no UI to change the seting.