Add a preference activity to choose the translation languages
This adds a list activity with all of the languages with a check box.
The user can select which translations to display. This sets the
selectedLanguages preference.

The article and select language activities have been modified to
register for preference changes on create and destroy instead of start
and stop. When the UI switches to another activity the old activity is
usually stopped so it is not enough to cause it to reload the state.
However to avoid redundantly reloading state for invisible activities,
it now queues a reload when the pref changes and only actually reloads
when the activity is started again.