Add a launch activity to skip the language select activity
The LanguagesActivity has been renamed to SelectLanguageActivity and a
new activity with the old name has been created to decide which
initial activity to show depending on the preferences. If the user has
previously already selected a language it will jump straight to the
search page for that language, otherwise it will default to the select
language activity.