LanguagesAdapter: Move list of languages to a separate helper class
The list of languages should never change throughout the lifetime of
the program so in order to aid sharing the list with other adapters it
is now stored in a shared global object. There is a new class to wrap
the array and it includes functions to search for a name given a code
by binary chopping the list.