Fix loading the article when there is a non-default font size
Previously if there was a non-default font size then it would try to
set the size on the views before the article has actually been loaded.
This patch changes it to delay setting the font size when there is a
reload queued. The font is always reset after the article is loaded
because the new views will again have the default size.

This patch also moves some initialisation out of loadIntendedArticle
into onCreate.