Cache the trie indices
Instead of each SearchActivity loading its own copy of the trie, they
are now cached in a static class so that they can be shared. The cache
only keeps the last 5 used indices so that if the user switches
between a lot of languages it won't use up too much memory.