SearchAdapter: Don't save the trie index
The trie is now loaded on-demand from the cache whenever a search is
initiatied rather than keeping it around forever. This should usually
hit the trie cache so it shouldn't cause any delays but it also has
the benefit that if the user opens a bunch of activities with many
different languages then the garbage collector will get a chance to
collect the tries. They will be loaded again when a new search is
started on the activity.