Don't mark a language as used if coming from a generic search
If MenuHelper.goSearch is called (ie, from the start activity or from
the generic search button) we don't want it to mark the language as
used because it's just using the default language so the user hasn't
really explicitly selected it. To make this work there is now an extra
string on the intent to specify if the language should be marked as
used.